+++ ECSSamples/Assets/HelloCube/2. IJobChunk/RotationSpeed_IJobChunk.cs
@@ -1,9 +1,0 @@
-using System;
-using Unity.Entities;
-
-// ReSharper disable once InconsistentNaming
-[Serializable]
-public struct RotationSpeed_IJobChunk : IComponentData
-{
-    public float RadiansPerSecond;
-}