+++ ECSSamples/Assets/Tests/ApiCompatibilityLevelTest.cs
@@ -1,3 +1,0 @@
-#if !NET_STANDARD_2_0
-#error Project is expected to have API Compatibility Level set to .NET Standard 2.0
-#endif