+++ ECSSamples/Assets/Use Case Samples/1. State Machine AI/StateTransitionVisualizationSystem.cs
@@ -11,7 +11,7 @@ using static Unity.Mathematics.math;
 #if !UNITY_DISABLE_MANAGED_COMPONENTS
 [UpdateInGroup(typeof(PresentationSystemGroup))]
 [UpdateBefore(typeof(RenderMeshSystemV2))]
-public class StateTransitionVisualizationSystem : SystemBase
+public partial class StateTransitionVisualizationSystem : SystemBase
 {
     private EntityQuery m_StateTransitionMaterialSingletonQuery;